item carrinho produto e usuario inserindo produtos no ecommerce

+++ src/main/java/com/ecommerce/controller/ProdutoController.java
@@ -1,7 +1,6 @@
 package com.ecommerce.controller;
 
 import com.ecommerce.model.Produto;
-import com.ecommerce.repository.ProdutoRepository;
 import com.ecommerce.service.ProdutoService;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,22 +15,31 @@ public class ProdutoController {
     private ProdutoService produtoService;
 
     @GetMapping("/")
-    public List<Produto> listar() {return this.produtoService.pegarTodosProduto();}
+    public List<Produto> listar() {
+        return this.produtoService.pegarTodosProduto();
+    }
 
     @GetMapping("/{id}")
-    public Produto buscarPorId(@PathVariable(value="id") int id){return this.produtoService.pegarProdutoPorId(id);}
+    public Produto buscarPorId(@PathVariable(value = "id") Long id) {
+        return this.produtoService.pegarProdutoPorId(id);
+    }
 
     @PostMapping("/")
     @Transactional
     public Produto salvar(@RequestBody Produto produto) {
         return this.produtoService.salvar(produto);
     }
+
     @PutMapping("/")
     @Transactional
-    public Produto alterar(@RequestBody Produto produto) {return this.produtoService.alterar(produto);}
+    public Produto alterar(@RequestBody Produto produto) {
+        return this.produtoService.alterar(produto);
+    }
 
     @DeleteMapping("/{id}")
     @Transactional
-    public void deletar(@PathVariable("id") long id) {this.produtoService.deletar(id);}
+    public void deletar(@PathVariable("id") long id) {
+        this.produtoService.deletar(id);
+    }
 
 }
